Decided to tinker a bit with the rescaled definition near the regime change from direct series to asymptotic strategies for the xv-rescaled methods and managed to squeeze out a couple more digits. Now the accuracies are pretty comparable and I think reasonably satisfying.
Also gave the testing file some TLC. Now just running the tests is
actually a very convenient way to check for worst-case accuracy
regressions, especially if you comment out the allocations test, which
takes a little while.

Bumping version now due to the improved accuracy. This is probably the
end of this little tinkering spell.