Cleaning up some paperscripts in preparation for (hopeful) clarity as scripts get updated in the revision process.
Unrelated, but I have added some code for hard-generating the UK
polynomials in a way that doesn't require Polynomials.jl and can further
reduce the dependency tree. As of now just dropping in that objects
breaks the test for no allocations in second derivatives, so it's not
ready yet. But it's on my mind.